+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Enhancement
 - show `fetching gas...` while fetching gas data
 - add beaconcha education in awesome link
 
+Refactor
+- move gas skills into separate file
+
 # 0.20 2012/12/17
 
 Feature